Move subcommand to separate file
