+++ cmd/topo/install_remoteproc_runtime.go
@@ -1,0 +1,57 @@
+package main
+
+import (
+	"os"
+
+	"github.com/arm/topo/internal/install"
+	"github.com/arm/topo/internal/output/printable"
+	"github.com/arm/topo/internal/output/templates"
+	"github.com/arm/topo/internal/ssh"
+	"github.com/spf13/cobra"
+)
+
+const remoteprocRepoURL = "arm/remoteproc-runtime"
+
+var installRemoteprocRuntimeCmd = &cobra.Command{
+	Use:   "remoteproc-runtime",
+	Short: "Install remoteproc-runtime and shim to a location on the target's PATH",
+	Long: `Install remoteproc-runtime and shim to a location on the target's PATH.
+
+Fetches binaries from https://github.com/` + remoteprocRepoURL + `
+Set GITHUB_TOKEN to authenticate with the GitHub API and avoid rate limits.
+
+Attempts to replace existing installations if found.
+Falls back to ~/bin if no suitable locations are automatically found.
+`,
+	Args: cobra.ExactArgs(0),
+	RunE: func(cmd *cobra.Command, args []string) error {
+		cmd.SilenceUsage = true
+		sshTarget, err := requireTarget(cmd)
+		if err != nil {
+			return err
+		}
+
+		outputFormat, err := resolveOutput(cmd)
+		if err != nil {
+			return err
+		}
+		p, err := installRemoteprocRuntime(ssh.Host(sshTarget))
+		if err != nil {
+			return err
+		}
+		return printable.Print(p, os.Stdout, outputFormat)
+	},
+}
+
+func installRemoteprocRuntime(targetHost ssh.Host) (printable.Printable, error) {
+	results, err := install.InstallBinariesFromGithubRelease(targetHost, remoteprocRepoURL, []string{"remoteproc-runtime", "containerd-shim-remoteproc-v1"})
+	if err != nil {
+		return nil, err
+	}
+	return templates.InstallResults(results), nil
+}
+
+func init() {
+	installCmd.AddCommand(installRemoteprocRuntimeCmd)
+	addTargetFlag(installRemoteprocRuntimeCmd)
+}
